test: Log command output at fail level if action failed
Log expected command failure at info level so that it is easier to
reason about the following flow logs, if any.

Always consider any stderr output as a failure and log it at fail
level.

Log stdout at fail level if the action failed, and at the debug level
otherwise.

This helps diagnoze test failures also when `--debug` is not on.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>